API: move phantoms to odl.phantoms, closes #463
Also did the following fixed to phantoms

* Optimized 2d ellipse phantoms in same way as 3d
* Fixed bug with submarined in non centered geometries
