+++ examples/solvers/chambolle_pock_tomography.py
@@ -64,11 +64,11 @@ ray_trafo = odl.tomo.RayTransform(reco_space, geometry, impl=impl)
 
 
 # Create phantom
-discr_phantom = odl.util.shepp_logan(reco_space, modified=True)
+discr_phantom = odl.phantom.shepp_logan(reco_space, modified=True)
 
 # Create sinogram of forward projected phantom with noise
 data = ray_trafo(discr_phantom)
-data += odl.util.white_noise(ray_trafo.range) * np.mean(data) * 0.1
+data += odl.phantom.white_noise(ray_trafo.range) * np.mean(data) * 0.1
 
 
 # --- Set up the inverse problem --- #
