+++ examples/solvers/conjugate_gradient_tomography.py
@@ -62,11 +62,11 @@ ray_trafo = odl.tomo.RayTransform(reco_space, geometry, impl=impl)
 
 
 # Create phantom
-discr_phantom = odl.util.shepp_logan(reco_space, modified=True)
+discr_phantom = odl.phantom.shepp_logan(reco_space, modified=True)
 
 # Create sinogram of forward projected phantom with noise
 data = ray_trafo(discr_phantom)
-data += odl.util.white_noise(ray_trafo.range) * np.mean(data) * 0.1
+data += odl.phantom.white_noise(ray_trafo.range) * np.mean(data) * 0.1
 
 # Optionally pass partial to the solver to display intermediate results
 partial = (odl.solvers.PrintIterationPartial() &
