+++ examples/tomo/ray_trafo_circular_cone.py
@@ -45,7 +45,7 @@ geometry = odl.tomo.CircularConeFlatGeometry(
 ray_trafo = odl.tomo.RayTransform(reco_space, geometry, impl='astra_cuda')
 
 # Create a discrete Shepp-Logan phantom (modified version)
-phantom = odl.util.phantom.shepp_logan(reco_space, True)
+phantom = odl.phantom.shepp_logan(reco_space, True)
 
 # Create projection data by calling the ray transform on the phantom
 proj_data = ray_trafo(phantom)
