+++ examples/tomo/ray_trafo_fanflat.py
@@ -44,7 +44,7 @@ geometry = odl.tomo.FanFlatGeometry(angle_partition, detector_partition,
 ray_trafo = odl.tomo.RayTransform(reco_space, geometry, impl='astra_cuda')
 
 # Create a discrete Shepp-Logan phantom (modified version)
-phantom = odl.util.phantom.shepp_logan(reco_space, modified=True)
+phantom = odl.phantom.shepp_logan(reco_space, modified=True)
 
 # Create projection data by calling the ray transform on the phantom
 proj_data = ray_trafo(phantom)
