+++ setup.py
@@ -89,7 +89,7 @@ Features
 setup(
     name='odl',
 
-    version='0.2.3',
+    version='0.2.4',
 
     description='Operator Discretization Library',
     long_description=long_description,
